+++ src/by/nik/traning/entity/Festival.java
@@ -4,7 +4,7 @@ import by.nik.traning.EventException;
 
 public class Festival extends Event {
 
-    private Long id;
+    private Integer id;
     private String description;
     private String date;
     private String place;
@@ -13,7 +13,7 @@ public class Festival extends Event {
 
     public Festival(){}
 
-	public Festival(Long id, String description, String date, String place, Integer countPlace, Integer countUser) {
+	public Festival(Integer id, String description, String date, String place, Integer countPlace, Integer countUser) {
 		super(id, description, date, place, countPlace, countUser);
 	}
 
@@ -28,10 +28,10 @@ public class Festival extends Event {
 	public void setCountPlace(Integer countPlace) {
 		this.countPlace = countPlace;
 	}
-    public Long getId() {
+    public Integer getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 	public String getDescription() {
